projects/service-zuscale/requirements.txt
aiohttp
requests
redis
hiredis
aiobotocore
aiolimiter
pyyaml
orjson
